Separate QueryInfo from QueryRegenerateEvent this removes some useless and confusing APIs from query info.
